Add files via upload
Statistical Analysis of coffee consumption

The file (Coffee.xlsx) contains survey data with detailed description given in the attached doc (Desc.docx). The survey aims to understand coffee consumption pattern of households.

It also contains an R script file to understnad the workflow of the analysis and Documented analysis to understand in detail.

Specifically a few of the questions the survey may attempt to answer are as follows:-

Which brands of coffee are more popular? Given a brand, are all variants equally preferred?
What are the prices of different brands of coffee?
How frequently does a household buy coffee? How many packets of coffee are bought at a time?
What are the factors that have an impact on a household’s coffee purchase pattern? Does brand preference depend on household size? Does purchase depend on a person’s income or education level?
There may be many other related and relevant questions that a company may be interested in. But before we get to do that, we need to ‘prepare’ the data and do some simple checks. The list is endless. However, a starting point could be the following:

Are there any variables/columns you think will be eventually irrelevant for the study? Why?
Can you possibly identify some outliers?
For better analysis do you think a few brand variants needs to be clubbed together?
Can you do some probability estimates- Say what is the probability that a randomly chosen household belongs to the top two social levels? What is the probability that households belonging to the top two social levels buy more expensive coffee?